// src/generated/geo-search.ts
export interface components {
    schemas: {
// ...
        GeoSearchResponse: {
            /** @description The result parameters found within the requested search radius. */
            GeoSearchRS?: {
// ...
                GeoSearchResults?: {
// ...
                    GeoSearchResult?: {
                        /**
                         * @description Specifies the distance of the hotel from the given airport.
                         * @example 0.24
                         */
                        Distance?: number;
                        /**
                         * @description The direction of the specified location.
                         * @example N
                         */
                        Direction?: string;
                        /**
                         * @description The geographic latitude of the location.
                         * @example 32.880778
                         */
                        Latitude?: number;
                        /**
                         * @description The geographic longitude of the location.
                         * @example -96.961008
                         */
                        Longitude?: number;
                        /**
                         * @description The name of the location.
                         * @example MARRIOTT EXECUSTAY BEAVER CREEK
                         */
                        Name?: string;
                        /**
                         * @description The unique ID of the location.
                         * @example 86034
                         */
                        Id?: string;
                        /**
                         * @description The street address.
                         * @example 1000 Meadow Creek Drive
                         */
                        Street?: string;
                        /**
                         * @description The zip code of the location.
                         * @example 75038
                         */
                        Zip?: string;
                        /**
                         * @description The name of the city.
                         * @example Irvine
                         */
                        City?: string;
                        /**
                         * @description The two-letter state code.
                         * @example TX
                         */
                        State?: string;
                        /**
                         * @description The code of the country. The length of the code ranges from 2 to 50.
                         * @example US
                         */
                        Country?: string;
                        /**
                         * @description The type of the ConfidenceFactor ranges between 0 to 10.
                         * @example 1
                         */
                        ConfidenceFactor?: number;
                        /** @description Property filters; can be 'Name', 'Value'. */
                        Attributes?: {
                            /** @description Lists the property filters, e.g. 'Name' and 'Value'. */
                            Attribute?: {
                                /**
                                 * @description The specification when the property was filtered by name. Can only be 'LOCALAREA'.
                                 * @example LOCALAREA
                                 */
                                Name?: string;
                                /**
                                 * @description The unique ID of the attribute.
                                 * @example 900011743
                                 */
                                Id?: string;
                                /**
                                 * @description The specification when the response includes only locations inside a specified area when 'LOCALAREA' is present in 'Name'.
                                 * @example Carrier
                                 */
                                Value?: string;
                            }[];
                        };
                    }[];
                };
            };
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}